fix (provider-utils): detect failed fetch in browser environments
