+++ .changeset/small-deers-cover.md
@@ -1,0 +1,5 @@
+---
+'@ai-sdk/provider-utils': patch
+---
+
+fix (provider-utils): detect failed fetch in browser environments
